Exception handling

+++ src/main/java/com/fkbinho/commerce/Services/ProductService.java
@@ -1,15 +1,19 @@
 package com.fkbinho.commerce.Services;
 
 import com.fkbinho.commerce.Repositories.ProductRepository;
+import com.fkbinho.commerce.Services.exceptions.DatabaseException;
+import com.fkbinho.commerce.Services.exceptions.ResourceNotFoundException;
 import com.fkbinho.commerce.dto.ProductDTO;
 import com.fkbinho.commerce.entities.Product;
+import jakarta.persistence.EntityNotFoundException;
+import org.hibernate.exception.DataException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 public class ProductService {
@@ -19,8 +23,10 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = productRepository.findById(id).get();
-        return new ProductDTO(product);
+            Product product = productRepository.findById(id).orElseThrow(
+                    () -> new ResourceNotFoundException("Recurso não encontrado.")
+            );
+            return new ProductDTO(product);
     }
 
     @Transactional(readOnly = true)
@@ -39,15 +45,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product entity = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, entity);
-        entity = productRepository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, entity);
+            entity = productRepository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado.");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha na integridade referencial.");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product entity) {
